Implementation of STEP C on boolean/comparison expression
Lazy evaluation for boolean expression has been implemented using control flow with two function codeGenExprOnExpr which is called only on the highest node corresponding to the expression and take care of storing the result, the other is called recursively on children and take care of generating smaller expressions and jumps.

In the same way, comparison expression can be mixed in boolean expression and thus override these two methods.